Reset the CSS styling for the WeatherApp
Going to work on making the styling more consistent and clean.

This commit removes e.g. the vertical effect on the WeatherCards in Weather-
Display; where each individual card was offset on the y axis dependent on
the temperature it was displaying. This feature will likely be re-added as
the work of cleaning up the styling progresses.